+++ workspaces/mcp-integrations/plugins/techdocs-mcp-tool/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @red-hat-developer-hub/backstage-plugin-techdocs-mcp-tool
 
+## 0.2.1
+
+### Patch Changes
+
+- 0ba2481: Make tag optional parameter in the output schema of the MCP tools
+
 ## 0.2.0
 
 ### Minor Changes